move input and footer tests back to snap-ui-renderer
